consolidate status
